requirements.txt
Django==5.1.7
graphene-django==3.2.2
